streamlit_app/requirements.txt
streamlit==1.24.1
pandas
SQLAlchemy
psycopg2-binary
plotly
requests
python-dotenv
pyyaml
